Mount the host kernel's seccomp.h header to the test container.
The seccomp.h header (/usr/include/linux/seccomp.h) in the test container
may not always match that of the host kernel, leading to problems when
building the libseccomp C library inside the test container. This
library is used by libseccomp-golang which is in turn used by
Sysbox, causing problems for Sysbox.

For example, I found out that Sysbox was not working correctly when libseccomp
was built inside the container image on my Ubuntu focal host. But it was working
correctly when libseccomp was build at host level.

After debugging, I found out that a recent change in the Ubuntu focal kernel
updated the definition of symbol SECCOMP_IOCTL_NOTIF_ID_VALID in the kernel:

"
commit a11e5a541ab1f7ac4216783f032017983f582b95
Author: Kees Cook <keescook@chromium.org>
Date:   Mon Jun 15 15:42:46 2020 -0700

    seccomp: Fix ioctl number for SECCOMP_IOCTL_NOTIF_ID_VALID

...

-#define SECCOMP_IOCTL_NOTIF_ID_VALID   SECCOMP_IOR(2, __u64)
+#define SECCOMP_IOCTL_NOTIF_ID_VALID   SECCOMP_IOW(2, __u64)
"

This change went into the Ubuntu focal 5.4.0-47 version, and the seccomp.h
header file inside the sysbox test container has the updated definition
(because it's based on the ubuntu:latest docker image).

However, my host has an earlier kernel (5.4.0-45). As a result, the seccomp.h
with the updated definition in the test container does not work on my older
kernel.

To work-around this, I am mounting into the sysbox test container the
seccomp.h file from the host. This way, the file will match the host's kernel
definitions. I verfified that with this change, it's possible to build
libseccomp inside the test container and things work fine.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>