Made adjustment to integration-testcase to allow sysfs mount attributes to be properly parsed
As a consequence of SELinux's default activation in CentOS-8, most of the file-systems being mounted display 'seclabel' superblock attribute (see below). In this PR we are simply relaxing the parsing logic in one of our testcases to account for this case.

/ # cat /proc/$$/mountinfo
...
730 714 0:121 / /sys rw,nosuid,nodev,noexec,relatime - sysfs sysfs rw,seclabel
...

/ # mount | grep sysfs
sysfs on /sys type sysfs (rw,seclabel,nosuid,nodev,noexec,relatime)
/ #

Signed-off-by: Rodny Molina <rmolina@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>