In sysbox-in-docker scenarios adjust initialization logic to catch sysfs in RO mode
This is a workaround for [issue](https://github.com/containerd/containerd/issues/3221).

This issue is observed in sysbox-in-docker scenarios when a privileged container is created by K8s/docker-shim to host the sysbox environment. In certain cases (still to be determined, but outside Sysbox scope), /sys file-system is mounted as read-only, which prevents Sysbox's proper operation. The workaround offered here is to verify /sys's proper setup during Sysbox initialization, and proceed to remount it as read-write if needed.

A similar verification logic will be added to sysbox-mgr to address this issue in general scenarios (beyond sysbox-in-docker ones).

Signed-off-by: Rodny Molina <rmolina@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>