Sysbox build/test container: use the old definition for SECCOMP_NOTIF_ID_VALID
Recently, the definition of SECCOMP_NOTIF_ID_VALID changed in the
mainline kernel, from SECCOMP_IOR(2, __u64) to SECCOMP_IOW(2,
__u64). The change ocurred on 06/2020.

Some distros we support have picked it up in their latest releases or
kernels updates, and this change is reflected in the libc header
"/usr/include/linux/seccomp.h" of said distros. Sysbox uses that
header when building the libseccomp library.

The problem is that if Sysbox is built with the new definition for
SECCOMP_NOTIF_ID_VALID, it will not run correctly on machines
with kernels that have the old definition (i.e., the sysbox-fs log
will report seccomp time-of-check-time-of-use (TOCTOU) failures).

Fortunately the kernel change was backward compatible, so by using the
old definition for SECCOMP_NOTIF_ID_VALID when building Sysbox, we are
guaranteed it will work on kernels that have the old or new
definition.

This commit ensures that inside the Sysbox test/build containers, we
always use the old definition of SECCOMP_NOTIF_ID_VALID in
"/usr/include/linux/seccomp.h".

Signed-off-by: ctalledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>