Address process' reaping issues in sysbox-in-docker scenarios
There are various issues with the existing initialization approach utilized in sysbox-in-docker scenarios. Among them is the lack of reaping functionality to handle zombie processes, as well as the absence of signal-handling capabilities to alert child processes of actionable signals being received (e.g. sigterm handling would allow graceful-restart capabilities).

For this purpose we are installing s6-overlay, a feature-rich and lightweight process manager to fill the gaps of the previous approach.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>