Minor fix in sysbox startup script.
In a prior commit, we changed the way we invoke sysbox-fs and sysbox-mgr
in the sysbox startup script: rather than redirecting their output
to a log file via a shell redirect, we used the `--log` command
line option.

This led to the following transient error when starting sysbox:

Starting sysbox with test-specific options
grep: /var/log/sysbox-mgr.log: No such file or directory
grep: /var/log/sysbox-fs.log: No such file or directory

The error occurs because the sysbox startup script checks
for the presence of the sysbox log files before sysbox
has a chance to create them. This commit fixes this.

Signed-off-by: ctalledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>