Add network config options to the docker-cfg script.
This change modifies the docker-cfg script to remove the "--skip-network-config"
option and replace it with the "--bip" and "--default-address-pool" options.

The latter allow the caller to tell the docker-cfg script how it should
configure the IP addresses for the docker0 bridge and for any
user-defined bridges. This provides more flexibility to the user.

As part of this change, the testContainerInit script has also
been updated to use the newly added docker-cfg options to
configure the docker network inside the test container such
that it won't overlap with docker networks on the host or
docker networks in inner containers.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>