First update to sysbox test infra to support testing of sysbox-pods.
This change modifiest the sysbox test container so that crictl and CRI-O can run
inside of it. This way, we can write integration tests that setup pods with
crictl + CRI-O + sysbox.

I've only updated the ubuntu-focal test container image; I'll update the other
test container images in a future commit.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>