Relax seccomp-fd deallocation interval in sysbox-fs ITs [skip ci]
The `sysfs/seccomp-notif.bats` testcase is regularly failing when executed in a Fedora host. It's not clear if the reason for the failure is coz, in average, Fedora takes slightly more time to react to the tracee's dead than other distributions, or that alternatively, Fedora is slightly faster than other distros in iterating through the test-case instructions. In either case, what is clear is that there's no seccomp-fd leak in any of the supported distributions.

To address Fedora's failures i'm simply adding a 1-sec delay between the 'pre' and 'post' stages of these testcases.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>