Rename env var SHIFT_UIDS to SHIFT_ROOTFS_UIDS in scripts.
This renaming is necessary for accuracy, as SHIFT_UIDS means that the test infra
will setup the container's rootfs with root:root ownership, forcing Sysbox to
perform uid shifting on the rootfs.

Note that SHIFT_UIDS does not necessarily mean that bind mounts into the
container will require uid shifting. Whether they do or not depends on the
ownership of the directory being bind-mounted into the container. Different
tests set that ownership in different ways, regardless of the SHIFT_UIDS env
variable.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>