Fix tests scripts to wait for Sysbox to fully initialize.
The sysbox test scripts had a bug that caused them to not wait sufficient time
for sysbox to be ready when sysbox is started / restarted.

This was causing some tests (e.g., sysmgr/kill.bats) to fail
intermittently.

The issue was worsened by a recent commit to the sysbox manual
startup script that added a delay during restarts (to fix another
problem).

This commit fixes the sysbox tests scripts to wait sufficient
time for sysbox to be ready.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>