Re-enable the docker cli+eng in one pod test.
The test was previously skipped as it was starting docker inside the pod using a
'crictl exec' background command, and this does not work unless sysbox-fs is
started with the --seccomp-fd-release=cont-exit option.

However, it's easy to work-around this having the pod's container.json spec
start Docker itself, so we can re-instate the test.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>