Remove Makefile dependency on lsb_release.
The Sysbox Makefile uses the lsb_release command in a few places. However, in
some distros (e.g., fedora), the lsb_release command is not present by default
and installing it requires installing a larger package such as redhat-lsb-core.

This change replaces the use of lsb_release with equivalent parsing of the
"/etc/os-release" file.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>