Add work-around for docker-cp in k3s test.
For some reason "docker cp" is failing to copy to/from subdirs in the container
(possibly related to a bug in docker cp caused by rootfs ID-mapping by
Sysbox). This needs further investigation.  As a temporary workaround, update a
K3s test that relies on docker cp to move the file to the container's "/" and
copy from there.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>