Update sysbox-docker-cp to work with idmapped rootfs.
The sysbox-docker-cp script was failing to chown files properly
when Sysbox uses idmapped-mounts on the container's rootfs
(the common case since kernel 5.19). Update the script to deal
with this scenario.

Note: unfortunately this requires that when copying files from
the host to the container, the user invoke this script with
"sudo", as the script needs to write to the container rootfs
under "/var/lib/docker" which is restricted to root-only access.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>

Signed-off-by: Amin Sanati <amin.sanati32@gmail.com>